requirements.txt
pandas==1.4.3
python-dotenv==0.21.0
tweepy==4.10.0
